.history/project-root/services/inference_service/requirements_20250419211221.txt
fastapi>=0.111
uvicorn[standard]>=0.29
torch==2.2.2          # or whatever Colab used
torchvision==0.17.2
pillow>=10.2
